+++ packages/protocol-parser/src/common.ts
@@ -1,6 +1,11 @@
-import { Schema } from "./Schema";
+import { DynamicAbiType, StaticAbiType } from "@latticexyz/schema-type";
+
+export type Schema = {
+  readonly staticFields: readonly StaticAbiType[];
+  readonly dynamicFields: readonly DynamicAbiType[];
+};
 
 export type TableSchema = {
-  keySchema: Schema;
-  valueSchema: Schema;
+  readonly keySchema: Schema; // TODO: refine to set dynamicFields to []
+  readonly valueSchema: Schema;
 };
