+++ packages/protocol-parser/src/hexToTableSchema.ts
@@ -1,10 +1,10 @@
 import { Hex, sliceHex } from "viem";
 import { TableSchema } from "./common";
-import { Schema } from "./Schema";
+import { hexToSchema } from "./hexToSchema";
 
 export function hexToTableSchema(data: Hex): TableSchema {
-  const valueSchema = Schema.fromHex(sliceHex(data, 0, 32));
-  const keySchema = Schema.fromHex(sliceHex(data, 32, 64));
+  const valueSchema = hexToSchema(sliceHex(data, 0, 32));
+  const keySchema = hexToSchema(sliceHex(data, 32, 64));
   return {
     keySchema,
     valueSchema,
