+++ packages/protocol-parser/src/index.ts
@@ -5,4 +5,3 @@ export * from "./decodeStaticField";
 export * from "./errors";
 export * from "./hexToPackedCounter";
 export * from "./hexToTableSchema";
-export * from "./Schema";
