+++ packages/protocol-parser/src/staticDataLength.ts
@@ -1,0 +1,5 @@
+import { StaticAbiType, staticAbiTypeToByteLength } from "@latticexyz/schema-type";
+
+export function staticDataLength(staticFields: readonly StaticAbiType[]): number {
+  return staticFields.reduce((length, fieldType) => length + staticAbiTypeToByteLength[fieldType], 0);
+}
